+++ ot/gromov/_unbalanced.py
@@ -12,7 +12,7 @@ import warnings
 from functools import partial
 from ot.backend import get_backend
 from ot.utils import list_to_array, get_parameter_pair
-from ot.unbalanced import sinkhorn_unbalanced, mm_unbalanced, lbfgsb_unbalanced
+from ._utils import total_cost, local_cost, parameters_uot_l2, uot_solver
 
 
 def fused_unbalanced_cross_spaces_divergence(
@@ -23,13 +23,18 @@ def fused_unbalanced_cross_spaces_divergence(
         tol=1e-7, max_iter_ot=500, tol_ot=1e-7, method_sinkhorn="sinkhorn",
         log=False, verbose=False, **kwargs_solver):
 
-    r"""Compute the fused unbalanced cross-spaces divergence between two matrices.
+    r"""Compute the fused unbalanced cross-spaces divergence between two matrices equipped
+    with the distributions of rows and columns. We consider two cases of matrix:
+    - (Squared) similarity matrix in Gromov-Wasserstein setting,
+    whose rows and columns represent the samples.
+    - Arbitrary-size matrix in Co-Optimal Transport setting,
+    whose rows represent samples and columns represent corresponding features/dimensions.
 
     Return the sample and feature transport plans between
     :math:`(\mathbf{X}, \mathbf{w}_{xs}, \mathbf{w}_{xf})` and
     :math:`(\mathbf{Y}, \mathbf{w}_{ys}, \mathbf{w}_{yf})`.
 
-    The function solves the following problem:
+    The function solves the following problem using Block Coordiante Descent algorithm:
 
     .. math::
         \mathbf{Div} = \mathop{\arg \min}_{\mathbf{P}, \mathbf{Q}}
@@ -43,8 +48,8 @@ def fused_unbalanced_cross_spaces_divergence(
 
     Where :
 
-    - :math:`\mathbf{X}`: Data matrix in the source space
-    - :math:`\mathbf{Y}`: Data matrix in the target space
+    - :math:`\mathbf{X}`: Source input (arbitrary-size) matrix
+    - :math:`\mathbf{Y}`: Target input (arbitrary-size) matrix
     - :math:`\mathbf{M^{(s)}}`: Additional sample matrix
     - :math:`\mathbf{M^{(f)}}`: Additional feature matrix
     - :math:`\mathbf{w}_{xs}`: Distribution of the samples in the source space
@@ -53,7 +58,7 @@ def fused_unbalanced_cross_spaces_divergence(
     - :math:`\mathbf{w}_{yf}`: Distribution of the features in the target space
     - :math:`\mathbf{Div}`: Either Kullback-Leibler divergence or half-squared L2 norm.
     - :math:`\mathbf{Reg}`: Regularizer for sample and feature couplings.
-    We consider two types of regulizer:
+    We consider two types of regularizer:
         + Independent regularization used in unbalanced Co-Optimal Transport
         .. math::
             \mathbf{Reg}(\mathbf{P}, \mathbf{Q}) =
@@ -71,9 +76,9 @@ def fused_unbalanced_cross_spaces_divergence(
     Parameters
     ----------
     X : (n_sample_x, n_feature_x) array-like, float
-        First input matrix.
+        Source input matrix.
     Y : (n_sample_y, n_feature_y) array-like, float
-        Second input matrix.
+        Target input matrix.
     wx_samp : (n_sample_x, ) array-like, float, optional (default = None)
         Histogram assigned on rows (samples) of matrix X.
         Uniform distribution by default.
@@ -99,8 +104,8 @@ def fused_unbalanced_cross_spaces_divergence(
         reg_type = "joint": then use joint regularization for couplings.
         reg_type = "indepedent": then use independent regularization for couplings.
     divergence : string, optional (default = "kl")
-        If divergence = "kl", then D is the Kullback-Leibler divergence.
-        If divergence = "l2", then D is the half squared Euclidean norm.
+        If divergence = "kl", then Div is the Kullback-Leibler divergence.
+        If divergence = "l2", then Div is the half squared Euclidean norm.
     unbalanced_solver : string, optional (default = "scaling")
         Solver for the unbalanced OT subroutine.
         If divergence = "kl", then unbalanced_solver can be: "scaling", "mm", "lbfgsb"
@@ -162,208 +167,6 @@ def fused_unbalanced_cross_spaces_divergence(
                 Total cost.
     """
 
-    #############################
-    # Calculate D(pi, a \otimes b) for D = KL and squared L2.
-
-    # Calculate KL(pi, a \otimes b) using
-    # the marginal distributions pi1, pi2 of pi
-    def approx_shortcut_kl(pi, pi1, pi2, a, b):
-        """
-        Implement:
-        < pi, log pi / (a \otimes b) >
-        = <pi, log pi> - <pi1, log a> - <pi2, log b>.
-        """
-
-        res = nx.sum(pi * nx.log(pi + 1.0 * (pi == 0))) \
-            - nx.sum(pi1 * nx.log(a)) - nx.sum(pi2 * nx.log(b))
-        return res
-
-    def div(pi, pi1, pi2, a, b):
-        """
-        Calculate D(pi, a \otimes b).
-        """
-
-        if divergence == "kl":
-            res = approx_shortcut_kl(pi, pi1, pi2, a, b) \
-                - nx.sum(pi1) + nx.sum(a) * nx.sum(b)
-        elif divergence == "l2":
-            res = (nx.sum(pi**2) + nx.sum(a**2) * nx.sum(b**2)
-                   - 2 * nx.dot(a, pi @ b)) / 2
-        return res
-
-    #############################
-    # Support functions for KL and squared L2 between product measures:
-    # Calculate D(mu \otimes nu, alpha \otimes beta).
-    def approx_kl(p, q):
-        return nx.sum(p * nx.log(p + 1.0 * (p == 0))) - nx.sum(p * nx.log(q))
-
-    def kl(p, q):
-        return approx_kl(p, q) - nx.sum(p) + nx.sum(q)
-
-    def product_kl(mu, nu, alpha, beta):
-        """
-        Calculate the KL divergence between two product measures:
-        KL(mu \otimes nu, alpha \otimes beta) =
-        m_mu * KL(nu, beta) + m_nu * KL(mu, alpha) +
-        (m_mu - m_alpha) * (m_nu - m_beta)
-
-        Parameters
-        ----------
-        mu: vector or matrix
-        nu: vector or matrix
-        alpha: vector or matrix with the same size as mu
-        beta: vector or matrix with the same size as nu
-
-        Returns
-        ----------
-        KL divergence between two product measures
-        """
-
-        m_mu, m_nu = nx.sum(mu), nx.sum(nu)
-        m_alpha, m_beta = nx.sum(alpha), nx.sum(beta)
-        const = (m_mu - m_alpha) * (m_nu - m_beta)
-        res = m_nu * kl(mu, alpha) + m_mu * kl(nu, beta) + const
-
-        return res
-
-    def product_l2(mu, nu, alpha, beta):
-        """
-        norm = ||mu \otimes nu - alpha \otimes beta ||^2
-        = ||a||^2 ||b||^2 + ||mu||^2 ||nu||^2 - 2 < alpha, mu > < beta, nu >.
-        L2(mu \otimes nu, alpha \otimes beta) = norm / 2.
-        """
-        norm = nx.sum(alpha**2) * nx.sum(beta**2) \
-            - 2 * nx.sum(alpha * mu) * nx.sum(beta * nu) \
-            + nx.sum(mu**2) * nx.sum(nu**2)
-
-        return norm / 2
-
-    def product_div(mu, nu, alpha, beta):
-        if divergence == "kl":
-            return product_kl(mu, nu, alpha, beta)
-        elif divergence == "l2":
-            return product_l2(mu, nu, alpha, beta)
-
-    #############################
-    # Support functions for BCD schemes
-    def local_cost(data, pi, tuple_p, hyperparams):
-        """
-        Calculate cost matrix of the UOT subroutine
-        """
-
-        X_sqr, Y_sqr, X, Y, M = data
-        rho_x, rho_y, eps = hyperparams
-        a, b = tuple_p
-
-        pi1, pi2 = nx.sum(pi, 1), nx.sum(pi, 0)
-        A, B = X_sqr @ pi1, Y_sqr @ pi2
-        uot_cost = A[:, None] + B[None, :] - 2 * X @ pi @ Y.T
-        if M is not None:
-            uot_cost = uot_cost + M
-
-        if divergence == "kl":
-            if rho_x != float("inf") and rho_x != 0:
-                uot_cost = uot_cost + rho_x * approx_kl(pi1, a)
-            if rho_y != float("inf") and rho_y != 0:
-                uot_cost = uot_cost + rho_y * approx_kl(pi2, b)
-            if reg_type == "joint" and eps > 0:
-                uot_cost = uot_cost + eps * approx_shortcut_kl(pi, pi1, pi2, a, b)
-
-        return uot_cost
-
-    def total_cost(M_linear, data, tuple_pxy_samp,
-                   tuple_pxy_feat, pi_samp, pi_feat, hyperparams):
-
-        rho_x, rho_y, eps_samp, eps_feat = hyperparams
-        M_samp, M_feat = M_linear
-        px_samp, py_samp, pxy_samp = tuple_pxy_samp
-        px_feat, py_feat, pxy_feat = tuple_pxy_feat
-        X_sqr, Y_sqr, X, Y = data
-
-        pi1_samp, pi2_samp = nx.sum(pi_samp, 1), nx.sum(pi_samp, 0)
-        pi1_feat, pi2_feat = nx.sum(pi_feat, 1), nx.sum(pi_feat, 0)
-
-        A_sqr = nx.dot(X_sqr @ pi1_feat, pi1_samp)
-        B_sqr = nx.dot(Y_sqr @ pi2_feat, pi2_samp)
-        AB = (X @ pi_feat @ Y.T) * pi_samp
-        linear_cost = A_sqr + B_sqr - 2 * nx.sum(AB)
-
-        if linear_cost < 0:
-            warnings.warn("The linear cost is negative: {}".format(linear_cost))
-
-        ucoot_cost = linear_cost
-        if M_samp is not None:
-            ucoot_cost = ucoot_cost + nx.sum(pi_samp * M_samp)
-        if M_feat is not None:
-            ucoot_cost = ucoot_cost + nx.sum(pi_feat * M_feat)
-
-        if rho_x != float("inf") and rho_x != 0:
-            ucoot_cost = ucoot_cost + \
-                rho_x * product_div(pi1_samp, pi1_feat, px_samp, px_feat)
-        if rho_y != float("inf") and rho_y != 0:
-            ucoot_cost = ucoot_cost + \
-                rho_y * product_div(pi2_samp, pi2_feat, py_samp, py_feat)
-
-        if reg_type == "joint" and eps_samp != 0:
-            div_cost = product_div(pi_samp, pi_feat, pxy_samp, pxy_feat)
-            ucoot_cost = ucoot_cost + eps_samp * div_cost
-        elif reg_type == "independent":
-            if eps_samp != 0:
-                div_samp = div(pi_samp, pi1_samp, pi2_samp, px_samp, py_samp)
-                ucoot_cost = ucoot_cost + eps_samp * div_samp
-            if eps_feat != 0:
-                div_feat = div(pi_feat, pi1_feat, pi2_feat, px_feat, py_feat)
-                ucoot_cost = ucoot_cost + eps_feat * div_feat
-
-        return linear_cost, ucoot_cost
-
-    # Support functions for squared L2 norm
-    def parameters_uot_l2(pi, tuple_weights, hyperparams):
-        """Compute parameters of the L2 loss."""
-
-        rho_x, rho_y, eps = hyperparams
-        wx, wy, wxy = tuple_weights
-
-        pi1, pi2 = nx.sum(pi, 1), nx.sum(pi, 0)
-        l2_pi1, l2_pi2, l2_pi = nx.sum(pi1**2), nx.sum(pi2**2), nx.sum(pi**2)
-
-        weighted_wx = wx * nx.sum(pi1 * wx) / l2_pi1
-        weighted_wy = wy * nx.sum(pi2 * wy) / l2_pi2
-        weighted_wxy = wxy * nx.sum(pi * wxy) / l2_pi if reg_type == "joint" else wxy
-        weighted_w = (weighted_wx, weighted_wy, weighted_wxy)
-
-        new_rho = (rho_x * l2_pi1, rho_y * l2_pi2)
-        new_eps = eps * l2_pi if reg_type == "joint" else eps
-
-        return weighted_w, new_rho, new_eps
-
-    # UOT solver for KL and squared L2
-    def uot_solver(wx, wy, wxy, cost, eps, rho, init_pi, init_duals):
-        if unbalanced_solver == "scaling":
-            pi, log = sinkhorn_unbalanced(
-                a=wx, b=wy, M=cost, reg=eps, reg_m=rho, reg_type="kl",
-                warmstart=init_duals, method=method_sinkhorn,
-                numItermax=max_iter_ot, stopThr=tol_ot, verbose=False, log=True)
-            duals = (log['logu'], log['logv'])
-
-        elif unbalanced_solver == "mm":
-            pi = mm_unbalanced(a=wx, b=wy, M=cost, reg_m=rho,
-                               c=wxy, reg=eps, div=divergence,
-                               G0=init_pi, numItermax=max_iter_ot,
-                               stopThr=tol_ot, verbose=False, log=False)
-            duals = (None, None)
-
-        elif unbalanced_solver == "lbfgsb":
-            pi = lbfgsb_unbalanced(a=wx, b=wy, M=cost, reg=eps, reg_m=rho,
-                                   c=wxy, reg_div=divergence,
-                                   regm_div=divergence,
-                                   G0=init_pi, numItermax=max_iter_ot,
-                                   stopThr=tol_ot, method='L-BFGS-B',
-                                   verbose=False, log=False)
-            duals = (None, None)
-
-        return pi, duals
-
     # MAIN FUNCTION
 
     if reg_type not in ["joint", "independent"]:
@@ -372,9 +175,6 @@ def fused_unbalanced_cross_spaces_divergence(
         raise (NotImplementedError('Unknown divergence="{}"'.format(divergence)))
     if unbalanced_solver not in ["scaling", "mm", "lbfgsb"]:
         raise (NotImplementedError('Unknown method="{}"'.format(unbalanced_solver)))
-
-    X, Y = list_to_array(X, Y)
-    nx = get_backend(X, Y)
 
     # hyperparameters
     alpha_samp, alpha_feat = get_parameter_pair(alpha)
@@ -390,6 +190,32 @@ def fused_unbalanced_cross_spaces_divergence(
         warnings.warn("Scaling algorithm does not support unregularized problem. \
                       Solver is set to 'mm'.")
         unbalanced_solver = "mm"
+
+    if init_pi is None:
+        pi_samp, pi_feat = None, None
+    else:
+        pi_samp, pi_feat = init_pi
+
+    if init_duals is None:
+        init_duals = (None, None)
+    duals_samp, duals_feat = init_duals
+
+    arr = [X, Y]
+
+    for item in [wx_samp, wx_feat, wy_samp, wy_feat,
+                 M_samp, M_feat, pi_samp, pi_feat]:
+        if item is not None:
+            arr.append(list_to_array(item))
+
+    for tuple in [duals_samp, duals_feat]:
+        if tuple is not None:
+            d1, d2 = duals_feat
+            if d1 is not None:
+                arr.append(list_to_array(d1))
+            if d2 is not None:
+                arr.append(list_to_array(d2))
+
+    nx = get_backend(*arr)
 
     # constant input variables
     if M_samp is None or alpha_samp == 0:
@@ -417,15 +243,9 @@ def fused_unbalanced_cross_spaces_divergence(
     wxy_feat = wx_feat[:, None] * wy_feat[None, :]
 
     # initialize coupling and dual vectors
-    if init_pi is None:
-        init_pi = (None, None)
-    pi_samp, pi_feat = init_pi
     pi_samp = wxy_samp if pi_samp is None else pi_samp
     pi_feat = wxy_feat if pi_feat is None else pi_feat
 
-    if init_duals is None:
-        init_duals = (None, None)
-    duals_samp, duals_feat = init_duals
     if unbalanced_solver == "scaling":
         if duals_samp is None:
             duals_samp = (nx.zeros(nx_samp, type_as=X),
@@ -439,20 +259,42 @@ def fused_unbalanced_cross_spaces_divergence(
     local_cost_samp = partial(local_cost,
                               data=(X_sqr, Y_sqr, X, Y, M_samp),
                               tuple_p=(wx_feat, wy_feat),
-                              hyperparams=(rho_x, rho_y, eps_feat))
+                              hyperparams=(rho_x, rho_y, eps_feat),
+                              divergence=divergence,
+                              reg_type=reg_type,
+                              nx=nx)
+
     local_cost_feat = partial(local_cost,
                               data=(X_sqr.T, Y_sqr.T, X.T, Y.T, M_feat),
                               tuple_p=(wx_samp, wy_samp),
-                              hyperparams=(rho_x, rho_y, eps_samp))
+                              hyperparams=(rho_x, rho_y, eps_samp),
+                              divergence=divergence,
+                              reg_type=reg_type,
+                              nx=nx)
+
     parameters_uot_l2_samp = partial(
         parameters_uot_l2,
         tuple_weights=(wx_samp, wy_samp, wxy_samp),
-        hyperparams=(rho_x, rho_y, eps_samp)
+        hyperparams=(rho_x, rho_y, eps_samp),
+        reg_type=reg_type,
+        nx=nx
     )
+
     parameters_uot_l2_feat = partial(
         parameters_uot_l2,
         tuple_weights=(wx_feat, wy_feat, wxy_feat),
-        hyperparams=(rho_x, rho_y, eps_feat)
+        hyperparams=(rho_x, rho_y, eps_feat),
+        reg_type=reg_type,
+        nx=nx
+    )
+
+    solver = partial(
+        uot_solver,
+        divergence=divergence,
+        unbalanced_solver=unbalanced_solver,
+        method_sinkhorn=method_sinkhorn,
+        max_iter_ot=max_iter_ot,
+        tol_ot=tol_ot
     )
 
     # initialize log
@@ -469,13 +311,13 @@ def fused_unbalanced_cross_spaces_divergence(
         if divergence == "kl":
             new_rho = (rho_x * mass, rho_y * mass)
             new_eps = mass * eps_feat if reg_type == "joint" else eps_feat
-            pi_feat, duals_feat = uot_solver(wx_feat, wy_feat, wxy_feat, uot_cost,
-                                             new_eps, new_rho, pi_feat, duals_feat)
+            pi_feat, duals_feat = solver(wx_feat, wy_feat, wxy_feat, uot_cost,
+                                         new_eps, new_rho, pi_feat, duals_feat)
         else:  # divergence == "l2"
             new_w, new_rho, new_eps = parameters_uot_l2_feat(pi_feat)
             new_wx, new_wy, new_wxy = new_w
-            pi_feat, duals_feat = uot_solver(new_wx, new_wy, new_wxy, uot_cost,
-                                             new_eps, new_rho, pi_feat, duals_feat)
+            pi_feat, duals_feat = solver(new_wx, new_wy, new_wxy, uot_cost,
+                                         new_eps, new_rho, pi_feat, duals_feat)
 
         if rescale_plan:
             pi_feat = nx.sqrt(mass / nx.sum(pi_feat)) * pi_feat
@@ -487,13 +329,13 @@ def fused_unbalanced_cross_spaces_divergence(
         if divergence == "kl":
             new_rho = (rho_x * mass, rho_y * mass)
             new_eps = mass * eps_feat if reg_type == "joint" else eps_feat
-            pi_samp, duals_samp = uot_solver(wx_samp, wy_samp, wxy_samp, uot_cost,
-                                             new_eps, new_rho, pi_samp, duals_samp)
+            pi_samp, duals_samp = solver(wx_samp, wy_samp, wxy_samp, uot_cost,
+                                         new_eps, new_rho, pi_samp, duals_samp)
         else:  # divergence == "l2"
             new_w, new_rho, new_eps = parameters_uot_l2_samp(pi_samp)
             new_wx, new_wy, new_wxy = new_w
-            pi_samp, duals_samp = uot_solver(new_wx, new_wy, new_wxy, uot_cost,
-                                             new_eps, new_rho, pi_samp, duals_samp)
+            pi_samp, duals_samp = solver(new_wx, new_wy, new_wxy, uot_cost,
+                                         new_eps, new_rho, pi_samp, duals_samp)
 
         if rescale_plan:
             pi_samp = nx.sqrt(mass / nx.sum(pi_samp)) * pi_samp  # shape nx x ny
@@ -519,7 +361,10 @@ def fused_unbalanced_cross_spaces_divergence(
             tuple_pxy_samp=(wx_samp, wy_samp, wxy_samp),
             tuple_pxy_feat=(wx_feat, wy_feat, wxy_feat),
             pi_samp=pi_samp, pi_feat=pi_feat,
-            hyperparams=(rho_x, rho_y, eps_samp, eps_feat)
+            hyperparams=(rho_x, rho_y, eps_samp, eps_feat),
+            divergence=divergence,
+            reg_type=reg_type,
+            nx=nx
         )
 
         dict_log["duals_sample"] = duals_samp
@@ -542,13 +387,13 @@ def unbalanced_co_optimal_transport(
         method_sinkhorn="sinkhorn", log=False, verbose=False,
         **kwargs_solve):
 
-    r"""Compute the unbalanced Co-Optimal Transport between two matrices.
+    r"""Compute the unbalanced Co-Optimal Transport between two arbitrary-size matrices.
 
     Return the sample and feature transport plans between
     :math:`(\mathbf{X}, \mathbf{w}_{xs}, \mathbf{w}_{xf})` and
     :math:`(\mathbf{Y}, \mathbf{w}_{ys}, \mathbf{w}_{yf})`.
 
-    The function solves the following problem:
+    The function solves the following problem using Block Coordiante Descent algorithm:
 
     .. math::
         \mathbf{UCOOT} = \mathop{\arg \min}_{\mathbf{P}, \mathbf{Q}}
@@ -563,8 +408,8 @@ def unbalanced_co_optimal_transport(
 
     Where :
 
-    - :math:`\mathbf{X}`: Data matrix in the source space
-    - :math:`\mathbf{Y}`: Data matrix in the target space
+    - :math:`\mathbf{X}`: Source input (arbitrary-size) matrix
+    - :math:`\mathbf{Y}`: Target input (arbitrary-size) matrix
     - :math:`\mathbf{M^{(s)}}`: Additional sample matrix
     - :math:`\mathbf{M^{(f)}}`: Additional feature matrix
     - :math:`\mathbf{w}_{xs}`: Distribution of the samples in the source space
@@ -579,9 +424,9 @@ def unbalanced_co_optimal_transport(
     Parameters
     ----------
     X : (n_sample_x, n_feature_x) array-like, float
-        First input matrix.
+        Source input matrix.
     Y : (n_sample_y, n_feature_y) array-like, float
-        Second input matrix.
+        Target input matrix.
     wx_samp : (n_sample_x, ) array-like, float, optional (default = None)
         Histogram assigned on rows (samples) of matrix X.
         Uniform distribution by default.
@@ -604,8 +449,8 @@ def unbalanced_co_optimal_transport(
         instead of Sinkhorn solver. If epsilon is scalar, then the same epsilon is applied to
         both regularization of sample and feature couplings.
     divergence : string, optional (default = "kl")
-        If divergence = "kl", then D is the Kullback-Leibler divergence.
-        If divergence = "l2", then D is the half squared Euclidean norm.
+        If divergence = "kl", then Div is the Kullback-Leibler divergence.
+        If divergence = "l2", then Div is the half squared Euclidean norm.
     unbalanced_solver : string, optional (default = "scaling")
         Solver for the unbalanced OT subroutine.
         If divergence = "kl", then unbalanced_solver can be: "scaling", "mm", "lbfgsb"
@@ -693,13 +538,13 @@ def unbalanced_co_optimal_transport2(
         method_sinkhorn="sinkhorn", log=False, verbose=False,
         **kwargs_solve):
 
-    r"""Compute the unbalanced Co-Optimal Transport between two matrices.
+    r"""Compute the unbalanced Co-Optimal Transport between two arbitrary-size matrices.
 
     Return the sample and feature transport plans between
     :math:`(\mathbf{X}, \mathbf{w}_{xs}, \mathbf{w}_{xf})` and
     :math:`(\mathbf{Y}, \mathbf{w}_{ys}, \mathbf{w}_{yf})`.
 
-    The function solves the following problem:
+    The function solves the following problem using Block Coordiante Descent algorithm:
 
     .. math::
         \mathbf{UCOOT} = \mathop{\arg \min}_{\mathbf{P}, \mathbf{Q}}
@@ -714,8 +559,8 @@ def unbalanced_co_optimal_transport2(
 
     Where :
 
-    - :math:`\mathbf{X}`: Data matrix in the source space
-    - :math:`\mathbf{Y}`: Data matrix in the target space
+    - :math:`\mathbf{X}`: Source input (arbitrary-size) matrix
+    - :math:`\mathbf{Y}`: Target input (arbitrary-size) matrix
     - :math:`\mathbf{M^{(s)}}`: Additional sample matrix
     - :math:`\mathbf{M^{(f)}}`: Additional feature matrix
     - :math:`\mathbf{w}_{xs}`: Distribution of the samples in the source space
@@ -733,9 +578,9 @@ def unbalanced_co_optimal_transport2(
     Parameters
     ----------
     X : (n_sample_x, n_feature_x) array-like, float
-        First input matrix.
+        Source input matrix.
     Y : (n_sample_y, n_feature_y) array-like, float
-        Second input matrix.
+        Target input matrix.
     wx_samp : (n_sample_x, ) array-like, float, optional (default = None)
         Histogram assigned on rows (samples) of matrix X.
         Uniform distribution by default.
@@ -758,8 +603,8 @@ def unbalanced_co_optimal_transport2(
         instead of Sinkhorn solver. If epsilon is scalar, then the same epsilon is applied to
         both regularization of sample and feature couplings.
     divergence : string, optional (default = "kl")
-        If divergence = "kl", then D is the Kullback-Leibler divergence.
-        If divergence = "l2", then D is the half squared Euclidean norm.
+        If divergence = "kl", then Div is the Kullback-Leibler divergence.
+        If divergence = "l2", then Div is the half squared Euclidean norm.
     unbalanced_solver : string, optional (default = "scaling")
         Solver for the unbalanced OT subroutine.
         If divergence = "kl", then unbalanced_solver can be: "scaling", "mm", "lbfgsb"
@@ -865,9 +710,9 @@ def unbalanced_co_optimal_transport2(
 
     # calculate subgradients
     gradX = 2 * X * (pi1_samp[:, None] * pi1_feat[None, :]) - \
-        2 * pi_samp @ Y @ pi_feat.T  # shape (nx_samp, nx_feat)
+        2 * nx.dot(nx.dot(pi_samp, Y), pi_feat.T)  # shape (nx_samp, nx_feat)
     gradY = 2 * Y * (pi2_samp[:, None] * pi2_feat[None, :]) - \
-        2 * pi_samp.T @ X @ pi_feat  # shape (ny_samp, ny_feat)
+        2 * nx.dot(nx.dot(pi_samp.T, X), pi_feat)  # shape (ny_samp, ny_feat)
 
     grad_wx_samp = rho_x * (m_wx_feat - m_feat * pi1_samp / wx_samp) + \
         eps_samp * (m_wy_samp - pi1_samp / wx_samp)
@@ -899,12 +744,12 @@ def fused_unbalanced_gromov_wasserstein(
         tol=1e-7, max_iter_ot=500, tol_ot=1e-7, method_sinkhorn="sinkhorn",
         log=False, verbose=False, **kwargs_solve):
 
-    r"""Compute the fused unbalanced Gromov-Wasserstein between two matrices.
+    r"""Compute the fused unbalanced Gromov-Wasserstein between two similarity matrices.
 
     Return the sample and feature transport plans between
     :math:`(\mathbf{C^X}, \mathbf{w_X})` and :math:`(\mathbf{C^Y}, \mathbf{w_Y})`.
 
-    The function solves the following problem:
+    The function solves the following problem using Block Coordiante Descent algorithm:
 
     .. math::
         \mathbf{FUGW} = \mathop{\arg \min}_{\mathbf{P}} &\quad \sum_{i,j,k,l}
@@ -916,9 +761,9 @@ def fused_unbalanced_gromov_wasserstein(
 
     Where :
 
-    - :math:`\mathbf{C^X}`: Data matrix in the source space
-    - :math:`\mathbf{C^Y}`: Data matrix in the target space
-    - :math:`\mathbf{M}`: Additional sample matrix
+    - :math:`\mathbf{C^X}`: Source similarity matrix
+    - :math:`\mathbf{C^Y}`: Target similarity matrix
+    - :math:`\mathbf{M}`: Sample matrix corresponding to the Wasserstein term
     - :math:`\mathbf{w_X}`: Distribution of the samples in the source space
     - :math:`\mathbf{w_Y}`: Distribution of the samples in the target space
     - :math:`\mathbf{Div}`: Either Kullback-Leibler divergence or half-squared L2 norm.
@@ -929,9 +774,9 @@ def fused_unbalanced_gromov_wasserstein(
     Parameters
     ----------
     Cx : (n_sample_x, n_feature_x) array-like, float
-        First input matrix.
+        Source similarity matrix.
     Cy : (n_sample_y, n_feature_y) array-like, float
-        Second input matrix.
+        Target similarity matrix.
     wx : (n_sample_x, ) array-like, float, optional (default = None)
         Histogram assigned on rows (samples) of matrix Cx.
         Uniform distribution by default.
@@ -948,8 +793,8 @@ def fused_unbalanced_gromov_wasserstein(
         instead of Sinkhorn solver. If epsilon is scalar, then the same epsilon is applied to
         both regularization of sample and feature couplings.
     divergence : string, optional (default = "kl")
-        If divergence = "kl", then D is the Kullback-Leibler divergence.
-        If divergence = "l2", then D is the half squared Euclidean norm.
+        If divergence = "kl", then Div is the Kullback-Leibler divergence.
+        If divergence = "l2", then Div is the half squared Euclidean norm.
     unbalanced_solver : string, optional (default = "scaling")
         Solver for the unbalanced OT subroutine.
         If divergence = "kl", then unbalanced_solver can be: "scaling", "mm", "lbfgsb"
@@ -1043,12 +888,12 @@ def fused_unbalanced_gromov_wasserstein2(
         tol=1e-7, max_iter_ot=500, tol_ot=1e-7, method_sinkhorn="sinkhorn",
         log=False, verbose=False, **kwargs_solve):
 
-    r"""Compute the fused unbalanced Gromov-Wasserstein between two matrices.
+    r"""Compute the fused unbalanced Gromov-Wasserstein between two similarity matrices.
 
     Return the sample and feature transport plans between
     :math:`(\mathbf{C^X}, \mathbf{w_X})` and :math:`(\mathbf{C^Y}, \mathbf{w_Y})`.
 
-    The function solves the following problem:
+    The function solves the following problem using Block Coordinate Descent algorithm:
 
     .. math::
         \mathbf{FUGW} = \mathop{\arg \min}_{\mathbf{P}} &\quad \sum_{i,j,k,l}
@@ -1060,9 +905,9 @@ def fused_unbalanced_gromov_wasserstein2(
 
     Where :
 
-    - :math:`\mathbf{C^X}`: Data matrix in the source space
-    - :math:`\mathbf{C^Y}`: Data matrix in the target space
-    - :math:`\mathbf{M}`: Additional sample matrix
+    - :math:`\mathbf{C^X}`: Source similarity matrix
+    - :math:`\mathbf{C^Y}`: Target similarity matrix
+    - :math:`\mathbf{M}`: Sample matrix corresponding to the Wasserstein term
     - :math:`\mathbf{w_X}`: Distribution of the samples in the source space
     - :math:`\mathbf{w_Y}`: Distribution of the samples in the target space
     - :math:`\mathbf{Div}`: Either Kullback-Leibler divergence or half-squared L2 norm.
@@ -1076,9 +921,9 @@ def fused_unbalanced_gromov_wasserstein2(
     Parameters
     ----------
     Cx : (n_sample_x, n_feature_x) array-like, float
-        First input matrix.
+        Source similarity matrix.
     Cy : (n_sample_y, n_feature_y) array-like, float
-        Second input matrix.
+        Target similarity matrix.
     wx : (n_sample_x, ) array-like, float, optional (default = None)
         Histogram assigned on rows (samples) of matrix Cx.
         Uniform distribution by default.
@@ -1095,8 +940,8 @@ def fused_unbalanced_gromov_wasserstein2(
         instead of Sinkhorn solver. If epsilon is scalar, then the same epsilon is applied to
         both regularization of sample and feature couplings.
     divergence : string, optional (default = "kl")
-        If divergence = "kl", then D is the Kullback-Leibler divergence.
-        If divergence = "l2", then D is the half squared Euclidean norm.
+        If divergence = "kl", then Div is the Kullback-Leibler divergence.
+        If divergence = "l2", then Div is the half squared Euclidean norm.
     unbalanced_solver : string, optional (default = "scaling")
         Solver for the unbalanced OT subroutine.
         If divergence = "kl", then unbalanced_solver can be: "scaling", "mm", "lbfgsb"
@@ -1186,9 +1031,9 @@ def fused_unbalanced_gromov_wasserstein2(
 
     # calculate subgradients
     gradX = 2 * Cx * (pi1_samp[:, None] * pi1_feat[None, :]) - \
-        2 * pi_samp @ Cy @ pi_feat.T  # shape (nx_samp, nx_feat)
+        2 * nx.dot(nx.dot(pi_samp, Cy), pi_feat.T)  # shape (nx_samp, nx_feat)
     gradY = 2 * Cy * (pi2_samp[:, None] * pi2_feat[None, :]) - \
-        2 * pi_samp.T @ Cx @ pi_feat  # shape (ny_samp, ny_feat)
+        2 * nx.dot(nx.dot(pi_samp.T, Cx), pi_feat)  # shape (ny_samp, ny_feat)
 
     gradM = alpha / 2 * (pi_samp + pi_feat)