+++ test/test_fugw.py
@@ -12,77 +12,7 @@ import pytest
 from ot.gromov._unbalanced import fused_unbalanced_gromov_wasserstein, fused_unbalanced_gromov_wasserstein2
 
 
-# @pytest.mark.parametrize("unbalanced_solver, divergence, eps", itertools.product(["scaling", "mm", "lbfgsb"], ["l2", "kl"], [0, 1e-2]))
-# def test_unbalanced_gromov_wasserstein(nx, unbalanced_solver, divergence, eps):
-#     n_samples = 5  # nb samples
-
-#     mu_s = np.array([0, 0])
-#     cov_s = np.array([[1, 0], [0, 1]])
-
-#     p = ot.unif(n_samples)
-#     q = ot.unif(n_samples)
-#     G0 = p[:, None] * q[None, :]
-
-#     xs = ot.datasets.make_2D_samples_gauss(n_samples, mu_s, cov_s, random_state=4)
-#     xt = xs[::-1].copy()
-
-#     C1 = ot.dist(xs, xs)
-#     C2 = ot.dist(xt, xt)
-#     C1 /= C1.max()
-#     C2 /= C2.max()
-
-#     C1b, C2b, pb, qb, G0b = nx.from_numpy(C1, C2, p, q, G0)
-
-#     reg_m = (10, 100)
-#     max_iter_ot = 1000
-#     max_iter = 1000
-#     tol = 1e-6
-#     tol_ot = 1e-6
-
-#     # test couplings
-#     pi_sample, pi_feature = fused_unbalanced_gromov_wasserstein(
-#         C1, C2, wx=p, wy=q, reg_marginals=reg_m, epsilon=eps,
-#         divergence=divergence, unbalanced_solver=unbalanced_solver,
-#         alpha=0, M=None, init_duals=None, init_pi=G0, max_iter=max_iter,
-#         tol=tol, max_iter_ot=max_iter_ot, tol_ot=tol_ot,
-#         method_sinkhorn="sinkhorn", log=False, verbose=False
-#     )
-
-#     pi_sample_nx, pi_feature_nx = fused_unbalanced_gromov_wasserstein(
-#         C1b, C2b, wx=pb, wy=qb, reg_marginals=reg_m, epsilon=eps,
-#         divergence=divergence, unbalanced_solver=unbalanced_solver,
-#         alpha=0, M=None, init_duals=None, init_pi=G0b, max_iter=max_iter,
-#         tol=tol, max_iter_ot=max_iter_ot, tol_ot=tol_ot,
-#         method_sinkhorn="sinkhorn", log=False, verbose=False
-#     )
-#     pi_sample_nx = nx.to_numpy(pi_sample_nx)
-#     pi_feature_nx = nx.to_numpy(pi_feature_nx)
-
-#     np.testing.assert_allclose(pi_sample, pi_sample_nx, atol=1e-03)
-#     np.testing.assert_allclose(pi_feature, pi_feature_nx, atol=1e-02)
-
-#     # test divergence
-
-#     fugw, log = fused_unbalanced_gromov_wasserstein2(
-#         C1, C2, wx=p, wy=q, reg_marginals=reg_m, epsilon=eps,
-#         divergence=divergence, unbalanced_solver=unbalanced_solver,
-#         alpha=0, M=None, init_duals=None, init_pi=G0, max_iter=max_iter,
-#         tol=tol, max_iter_ot=max_iter_ot, tol_ot=tol_ot,
-#         method_sinkhorn="sinkhorn", log=True, verbose=False
-#     )
-
-#     fugw_nx, log_nx = fused_unbalanced_gromov_wasserstein2(
-#         C1b, C2b, wx=pb, wy=qb, reg_marginals=reg_m, epsilon=eps,
-#         divergence=divergence, unbalanced_solver=unbalanced_solver,
-#         alpha=0, M=None, init_duals=None, init_pi=G0b, max_iter=max_iter,
-#         tol=tol, max_iter_ot=max_iter_ot, tol_ot=tol_ot,
-#         method_sinkhorn="sinkhorn", log=True, verbose=False
-#     )
-#     fugw_nx = nx.to_numpy(fugw_nx)
-
-#     np.testing.assert_allclose(fugw, fugw_nx, atol=1e-04)
-
-
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 @pytest.mark.parametrize("unbalanced_solver, divergence", itertools.product(["mm", "lbfgsb"], ["kl", "l2"]))
 def test_sanity(nx, unbalanced_solver, divergence):
     n_samples = 5  # nb samples
@@ -166,6 +96,7 @@ def test_sanity(nx, unbalanced_solver, divergence):
     np.testing.assert_allclose(fugw, 0, atol=1e-02)
 
 
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 @pytest.mark.parametrize("unbalanced_solver, divergence, eps", itertools.product(["scaling", "mm", "lbfgsb"], ["kl", "l2"], [0, 1e-2]))
 def test_init_plans(nx, unbalanced_solver, divergence, eps):
     n_samples = 5  # nb samples
@@ -240,6 +171,7 @@ def test_init_plans(nx, unbalanced_solver, divergence, eps):
     np.testing.assert_allclose(fugw, fugw_nx, atol=1e-08)
 
 
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 @pytest.mark.parametrize("unbalanced_solver, divergence, eps", itertools.product(["scaling", "mm", "lbfgsb"], ["kl", "l2"], [0, 1e-2]))
 def test_init_duals(nx, unbalanced_solver, divergence, eps):
     n_samples = 5  # nb samples
@@ -253,7 +185,8 @@ def test_init_duals(nx, unbalanced_solver, divergence, eps):
 
     p = ot.unif(n_samples)
     q = ot.unif(n_samples)
-    init_duals = (np.zeros(n_samples), np.zeros(n_samples))
+    dual1, dual2 = nx.from_numpy(np.zeros_like(p), np.zeros_like(q))
+    init_duals = (dual1, dual2)
 
     # linear part
     M_samp = np.ones((n_samples, n_samples))
@@ -263,6 +196,8 @@ def test_init_duals(nx, unbalanced_solver, divergence, eps):
     C2 = ot.dist(xt, xt)
     C1 /= C1.max()
     C2 /= C2.max()
+
+    C1, C2, p, q, M_samp = nx.from_numpy(C1, C2, p, q, M_samp)
 
     reg_m = (100, 50)
     alpha = 0.5
@@ -313,6 +248,7 @@ def test_init_duals(nx, unbalanced_solver, divergence, eps):
     np.testing.assert_allclose(fugw, fugw_nx, atol=1e-08)
 
 
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 @pytest.mark.parametrize("unbalanced_solver, divergence, eps", itertools.product(["scaling", "mm", "lbfgsb"], ["kl", "l2"], [0, 1e-2]))
 def test_reg_marginals(nx, unbalanced_solver, divergence, eps):
     n_samples = 5  # nb samples
@@ -393,6 +329,7 @@ def test_reg_marginals(nx, unbalanced_solver, divergence, eps):
         np.testing.assert_allclose(fugw, fugw_nx, atol=1e-08)
 
 
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 @pytest.mark.parametrize("unbalanced_solver, divergence, eps", itertools.product(["scaling", "mm", "lbfgsb"], ["kl", "l2"], [0, 1e-2]))
 def test_log(nx, unbalanced_solver, divergence, eps):
     n_samples = 5  # nb samples
@@ -466,6 +403,7 @@ def test_log(nx, unbalanced_solver, divergence, eps):
     np.testing.assert_allclose(fugw, fugw_nx, atol=1e-08)
 
 
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 @pytest.mark.parametrize("unbalanced_solver, divergence, eps", itertools.product(["scaling", "mm", "lbfgsb"], ["kl", "l2"], [0, 1e-2]))
 def test_marginals(nx, unbalanced_solver, divergence, eps):
     n_samples = 5  # nb samples
@@ -539,6 +477,7 @@ def test_marginals(nx, unbalanced_solver, divergence, eps):
     np.testing.assert_allclose(fugw, fugw_nx, atol=1e-08)
 
 
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 def test_raise_value_error(nx):
     n_samples = 5  # nb samples