+++ test/test_ucoot.py
@@ -12,6 +12,7 @@ import pytest
 from ot.gromov._unbalanced import unbalanced_co_optimal_transport, unbalanced_co_optimal_transport2
 
 
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 @pytest.mark.parametrize("unbalanced_solver, divergence", itertools.product(["mm", "lbfgsb"], ["kl", "l2"]))
 def test_sanity(nx, unbalanced_solver, divergence):
     n_samples = 5  # nb samples
@@ -89,6 +90,7 @@ def test_sanity(nx, unbalanced_solver, divergence):
     np.testing.assert_allclose(ucoot, 0, atol=1e-06)
 
 
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 @pytest.mark.parametrize("unbalanced_solver, divergence, eps", itertools.product(["scaling", "mm", "lbfgsb"], ["kl", "l2"], [0, 1]))
 def test_init_plans(nx, unbalanced_solver, divergence, eps):
     n_samples = 5  # nb samples
@@ -162,6 +164,7 @@ def test_init_plans(nx, unbalanced_solver, divergence, eps):
     np.testing.assert_allclose(ucoot, ucoot_nx, atol=1e-08)
 
 
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 @pytest.mark.parametrize("unbalanced_solver, divergence, eps", itertools.product(["scaling", "mm", "lbfgsb"], ["kl", "l2"], [0, 1]))
 def test_init_duals(nx, unbalanced_solver, divergence, eps):
     n_samples = 5  # nb samples
@@ -237,6 +240,7 @@ def test_init_duals(nx, unbalanced_solver, divergence, eps):
     np.testing.assert_allclose(ucoot, ucoot_nx, atol=1e-08)
 
 
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 @pytest.mark.parametrize("unbalanced_solver, divergence, eps", itertools.product(["scaling", "mm", "lbfgsb"], ["kl", "l2"], [0, 1e-2]))
 def test_linear_part(nx, unbalanced_solver, divergence, eps):
     n_samples = 5  # nb samples
@@ -315,6 +319,7 @@ def test_linear_part(nx, unbalanced_solver, divergence, eps):
     np.testing.assert_allclose(ucoot, ucoot_nx, atol=1e-08)
 
 
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 @pytest.mark.parametrize("unbalanced_solver, divergence, eps", itertools.product(["scaling", "mm", "lbfgsb"], ["kl", "l2"], [0, 1]))
 def test_reg_marginals(nx, unbalanced_solver, divergence, eps):
     n_samples = 5  # nb samples
@@ -395,6 +400,7 @@ def test_reg_marginals(nx, unbalanced_solver, divergence, eps):
         np.testing.assert_allclose(ucoot, ucoot_nx, atol=1e-08)
 
 
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 @pytest.mark.parametrize("unbalanced_solver, divergence, alpha", itertools.product(["scaling", "mm", "lbfgsb"], ["kl", "l2"], [0, 1]))
 def test_eps(nx, unbalanced_solver, divergence, alpha):
     n_samples = 5  # nb samples
@@ -476,6 +482,7 @@ def test_eps(nx, unbalanced_solver, divergence, alpha):
         np.testing.assert_allclose(ucoot, ucoot_nx, atol=1e-08)
 
 
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 @pytest.mark.parametrize("unbalanced_solver, divergence, eps", itertools.product(["scaling", "mm", "lbfgsb"], ["kl", "l2"], [0, 1e-2]))
 def test_alpha(nx, unbalanced_solver, divergence, eps):
     n_samples = 5  # nb samples
@@ -561,6 +568,7 @@ def test_alpha(nx, unbalanced_solver, divergence, eps):
         np.testing.assert_allclose(ucoot, ucoot_nx, atol=1e-08)
 
 
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 @pytest.mark.parametrize("unbalanced_solver, divergence, eps", itertools.product(["scaling", "mm", "lbfgsb"], ["kl", "l2"], [0, 1]))
 def test_log(nx, unbalanced_solver, divergence, eps):
     n_samples = 5  # nb samples
@@ -632,6 +640,7 @@ def test_log(nx, unbalanced_solver, divergence, eps):
     np.testing.assert_allclose(ucoot, ucoot_nx, atol=1e-08)
 
 
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 @pytest.mark.parametrize("unbalanced_solver, divergence, eps", itertools.product(["scaling", "mm", "lbfgsb"], ["kl", "l2"], [0, 1]))
 def test_marginals(nx, unbalanced_solver, divergence, eps):
     n_samples = 5  # nb samples
@@ -703,6 +712,7 @@ def test_marginals(nx, unbalanced_solver, divergence, eps):
     np.testing.assert_allclose(ucoot, ucoot_nx, atol=1e-08)
 
 
+@pytest.skip_backend("tf", reason="test very slow with tensorflow backend")
 def test_raise_value_error(nx):
     n_samples = 5  # nb samples